backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
psycopg2-binary
pandas
